Refactor calc_checksum to crc_checksum module
The calc_checksum is used by multiple modules, so should be in its
own module.

d_rats/crc_checksum: (new)
  calc_checksum moved from d_rats/ddt2.py

d_rats/ddt2.py:
  Remove calc_checksum function.
  Remove useless shebang and python2 import future.
  Renamed some variables from being marked as "unused" with a "_" prefix.

d_rats/wl2k.py:
  Change calc_checksum import.
  Remove useless shebang and python2 import future.

libexec/lzhuf.c:
  Fix spelling error.
